+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## master (XX.XX.XXXX)
+ + Add `automount` and `format` Parameters to `LKDev\HetznerCloud\Models\Volumes\Volumes` - `create`
+ + Add `created` property to `LKDev\HetznerCloud\Models\FloatingIps\FloatingIp`
+
 ## 1.6.0 (08.05.2019)
 + Setting the UserAgent is now possible with `LKDev\HetznerCloud\HetznerAPIClient` - `setUserAgent()`
 + Setting the Base URL is now possible with `LKDev\HetznerCloud\HetznerAPIClient` - `setBaseUrl()`